+++ boot/namespace.go
@@ -2,6 +2,7 @@ package boot
 
 import (
 	"context"
+	"strings"
 	"time"
 
 	"github.com/libp2p/go-libp2p/core/discovery"
@@ -15,33 +16,45 @@ type Namespace struct {
 }
 
 func (n Namespace) Network() string {
-	return "floodsub:" + n.Name
+	return n.Name
 }
 
-func (n Namespace) Advertise(ctx context.Context, ns string, opt ...discovery.Option) (time.Duration, error) {
-	logger := slog.Default().With(
+func (n Namespace) Advertise(ctx context.Context, net string, opt ...discovery.Option) (ttl time.Duration, err error) {
+	slog.Debug("advertising",
 		"ns", n.Name,
-		"net", ns)
+		"net", net)
 
-	ttl, err := n.Ambient.Advertise(ctx, ns, opt...)
-	if err != nil {
-		logger.With("error", err)
+	if strings.HasPrefix(net, "floodsub:") {
+		return trimPrefix{n.Bootstrap}.Advertise(ctx, net, opt...)
 	}
-	defer logger.Debug("advertised")
 
-	return ttl, err
+	return n.Ambient.Advertise(ctx, n.Name, opt...)
 }
 
-func (n Namespace) FindPeers(ctx context.Context, ns string, opt ...discovery.Option) (<-chan peer.AddrInfo, error) {
-	logger := slog.Default().With(
+func (n Namespace) FindPeers(ctx context.Context, net string, opt ...discovery.Option) (out <-chan peer.AddrInfo, err error) {
+	slog.Debug("finding peers",
 		"ns", n.Name,
-		"match", ns)
+		"net", net)
 
-	peers, err := n.Ambient.FindPeers(ctx, ns, opt...)
-	if err != nil {
-		logger.With("error", err)
+	if strings.HasPrefix(net, "floodsub:") {
+		return trimPrefix{n.Bootstrap}.FindPeers(ctx, net, opt...)
 	}
-	defer logger.Debug("crawled")
 
-	return peers, err
+	return n.Ambient.FindPeers(ctx, n.Name, opt...)
+}
+
+// Trims the "floodsub:" prefix from the namespace.  This is needed because
+// clients do not use pubsub, and will search for the exact namespace string.
+type trimPrefix struct {
+	discovery.Discovery
+}
+
+func (b trimPrefix) FindPeers(ctx context.Context, ns string, opt ...discovery.Option) (<-chan peer.AddrInfo, error) {
+	ns = strings.TrimPrefix(ns, "floodsub:")
+	return b.Discovery.FindPeers(ctx, ns, opt...)
+}
+
+func (b trimPrefix) Advertise(ctx context.Context, ns string, opt ...discovery.Option) (time.Duration, error) {
+	ns = strings.TrimPrefix(ns, "floodsub:")
+	return b.Discovery.Advertise(ctx, ns, opt...)
 }
